Split routing.rs into smaller modules
